pyproject.toml
[tool.poetry]
name = "multi-attach-mail"
version = "0.1.0"
description = ""
authors = ["Dr. Bastian Ebeling <230051+Barry1@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "multi_attach_mail" }]

[tool.poetry.dependencies]
aiosmtplib = ">=2"
pylyzer = "^0.0.51"
python = "^3.12"
PyYAML = ">6"
aiopath = "^0.7.7"

[tool.poetry.dependencies.valuefragments]
git = "https://github.com/Barry1/PyValueFragments"

[tool.poetry.group.dev.dependencies]
autopep8 = ">=2"
black = ">=23"
flynt = "^1.0.1"
mypy = "^1.1.1"
pylama = ">=8.4"
pylint = ">=2.17"
pyre-check = ">=0.9"
pyright = ">=1.1"
ruff = ">=0.3.7"
types-pyyaml = "^6.0.12.9"
sourcery = "^1.35.0"
vermin = "^1.6.0"

[tool.poetry.group.dev.dependencies.pydocstyle]
extras = ["toml"]
version = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylama]
verbose = true
#linters = "mccabe,mypy,pycodestyle,pydocstyle,pyflakes,pylint,isort"
linters = "mccabe,mypy,pycodestyle,pyflakes,pylint,isort"
